Reading some fintech posts suggest always using BigDecimal for currency rather than double/long/float

+++ src/main/java/com/rkcp/worthit/dto/PurchaseDecisionResponse.java
@@ -1,8 +1,10 @@
 package com.rkcp.worthit.dto;
 
+import java.math.BigDecimal;
+
 public record PurchaseDecisionResponse(
         String itemName,
-        double price,
-        double futureValue,
-        double interestGained,
+        BigDecimal price,
+        BigDecimal futureValue,
+        BigDecimal interestGained,
         String message) {}
